+++ cpp/include/algorithms.h
@@ -23,7 +23,7 @@
  * @Param[in] graph               cuGRAPH graph descriptor, should contain the connectivity information as an edge list (edge weights are not used for this algorithm).
  *                                The transposed adjacency list will be computed if not already present.
  * @Param[in] alpha               The damping factor alpha represents the probability to follow an outgoing edge, standard value is 0.85.
- Thus, 1.0-alpha is the probability to “teleport” to a random node. Alpha should be greater than 0.0 and strictly lower than 1.0.
+ Thus, 1.0-alpha is the probability to “teleport” to a random vertex. Alpha should be greater than 0.0 and strictly lower than 1.0.
  * @Param[in] has_guess           This parameter is used to notify cuGRAPH if it should use a user-provided initial guess. False means the user doesn't have a guess, in this case cuGRAPH will use a uniform vector set to 1/V.
  *                                If the value is True, cuGRAPH will read the pagerank parameter and use this as an initial guess.
  *                                The initial guess must not be the vector of 0s. Any value other than 1 or 0 is treated as an invalid value.
@@ -85,15 +85,15 @@ gdf_error gdf_grmat_gen(const char* argv,
                         gdf_column* val);
 
 /**
- * @Synopsis   Performs a breadth first search traversal of a graph starting from a node.
+ * @Synopsis   Performs a breadth first search traversal of a graph starting from a vertex.
  *
  * @Param[in] *graph                 cuGRAPH graph descriptor with a valid edgeList or adjList
  *
- * @Param[out] *distances            If set to a valid column, this is populated by distance of every vertex in the graph from the starting node
+ * @Param[out] *distances            If set to a valid column, this is populated by distance of every vertex in the graph from the starting vertex
  *
  * @Param[out] *predecessors         If set to a valid column, this is populated by bfs traversal predecessor of every vertex
  *
- * @Param[in] start_node             The starting node for breadth first search traversal
+ * @Param[in] start_vertex           The starting vertex for breadth first search traversal
  *
  * @Param[in] directed               Treat the input graph as directed
  *
@@ -103,7 +103,7 @@ gdf_error gdf_grmat_gen(const char* argv,
 gdf_error gdf_bfs(gdf_graph *graph,
                   gdf_column *distances,
                   gdf_column *predecessors,
-                  int start_node,
+                  int start_vertex,
                   bool directed);
 
 /**