+++ python/cugraph/jaccard/jaccard_wrapper.pyx
@@ -79,8 +79,8 @@ cpdef jaccard(input_graph, first=None, second=None):
     cdef gdf_column c_src_index_col
 
     if type(first) == cudf.dataframe.series.Series and type(second) == cudf.dataframe.series.Series:
-        resultSize = len(first)
-        result = cudf.Series(np.ones(resultSize, dtype=np.float32))
+        result_size = len(first)
+        result = cudf.Series(np.ones(result_size, dtype=np.float32))
         c_result_col = get_gdf_column_view(result)
         c_first_col = get_gdf_column_view(first)
         c_second_col = get_gdf_column_view(second)
@@ -97,18 +97,18 @@ cpdef jaccard(input_graph, first=None, second=None):
         return df
 
     elif first is None and second is None:
-        e = g.adjList.indices.size
-        result = cudf.Series(np.ones(e, dtype=np.float32), nan_as_null=False)
+        num_edges = input_graph.number_of_edges()
+        result = cudf.Series(np.ones(num_edges, dtype=np.float32), nan_as_null=False)
         c_result_col = get_gdf_column_view(result)
 
         err = gdf_jaccard(g, <gdf_column*> NULL, &c_result_col)
         cudf.bindings.cudf_cpp.check_gdf_error(err)
 
         dest_data = rmm.device_array_from_ptr(<uintptr_t> g.adjList.indices.data,
-                                            nelem=e,
+                                            nelem=num_edges,
                                             dtype=gdf_to_np_dtypes[g.adjList.indices.dtype])
         df = cudf.DataFrame()
-        df['source'] = cudf.Series(np.zeros(e, dtype=gdf_to_np_dtypes[g.adjList.indices.dtype]))
+        df['source'] = cudf.Series(np.zeros(num_edges, dtype=gdf_to_np_dtypes[g.adjList.indices.dtype]))
         c_src_index_col = get_gdf_column_view(df['source'])
         err = g.adjList.get_source_indices(&c_src_index_col);
         cudf.bindings.cudf_cpp.check_gdf_error(err)