+++ python/cugraph/overlap/test_overlap.py
@@ -55,10 +55,10 @@ def cugraph_call(M, first, second, edgevals=False):
 def intersection(a, b, M):
     count = 0
     for idx in range(M.indptr[a], M.indptr[a+1]):
-        a_node = M.indices[idx]
+        a_vertex = M.indices[idx]
         for inner_idx in range(M.indptr[b], M.indptr[b+1]):
-            b_node = M.indices[inner_idx]
-            if a_node == b_node:
+            b_vertex = M.indices[inner_idx]
+            if a_vertex == b_vertex:
                 count += 1
     return count